+++ lib/rerune_flutter_ota.dart
@@ -1,9 +1,7 @@
 export 'src/cache/cache_store.dart';
 export 'src/controller/ota_localization_controller.dart';
-export 'src/delegate/ota_localizations_delegate.dart';
-export 'src/localizations/ota_localizations.dart';
 export 'src/model/manifest.dart';
 export 'src/model/ota_error.dart';
 export 'src/model/update_result.dart';
 export 'src/policy/update_policy.dart';
-export 'src/widget/ota_localization_builder.dart';
+export 'src/widget/ota_typed_localization_builder.dart';